Rename the JS file
